Refactor HomeHeroBriefInformation component into its own file
